+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Package: iqnection-silverstripe-pages-pagebuilder
 # Change Log
 
-## [0.0.1]
+
+##
+- Added three column content
+
+##
 - Initial Release